Added the command that allows the user to add recipes to their recipe book, refactored the package the abstract class Command is in, added docstrings and some new methods to RecipeBookManager
