--- src/main/java/CLI/Pages/Command.java
@@ -1,22 +1,0 @@
-package main.java.CLI.Pages;
-
-import main.java.CLI.CommandLineInterface;
-
-public abstract class Command {
-    String name;
-    String description;
-    public Command(String name, String description) {
-        this.name = name;
-        this.description = description;
-    }
-    public String getName() {
-        return name;
-    }
-    public String getDescription() {
-        return description;
-    }
-    public abstract void runAction(CommandLineInterface CLI);
-}
-
-
-
