+++ src/main/java/CLI/Pages/Commands/Command.java
@@ -1,0 +1,22 @@
+package main.java.CLI.Pages.Commands;
+
+import main.java.CLI.CommandLineInterface;
+
+public abstract class Command {
+    String name;
+    String description;
+    public Command(String name, String description) {
+        this.name = name;
+        this.description = description;
+    }
+    public String getName() {
+        return name;
+    }
+    public String getDescription() {
+        return description;
+    }
+    public abstract void runAction(CommandLineInterface CLI);
+}
+
+
+
