+++ src/main/java/CLI/Pages/Commands/GetNewRecipeCommand.java
@@ -1,7 +1,6 @@
 package main.java.CLI.Pages.Commands;
 
 import main.java.CLI.CommandLineInterface;
-import main.java.CLI.Pages.Command;
 import main.java.Entities.Recipe;
 import main.java.Gateways.RecipeGateway;
 import main.java.Utilities.RecipeBookManager;
