+++ src/main/java/CLI/Pages/Commands/RemoveRecipeCommand.java
@@ -1,7 +1,6 @@
 package main.java.CLI.Pages.Commands;
 
 import main.java.CLI.CommandLineInterface;
-import main.java.CLI.Pages.Command;
 import main.java.Utilities.RecipeBookManager;
 
 /**
