+++ src/main/java/CLI/Pages/Commands/UpdatePreferencesCommand.java
@@ -1,7 +1,6 @@
 package main.java.CLI.Pages.Commands;
 
 import main.java.CLI.CommandLineInterface;
-import main.java.CLI.Pages.Command;
 
 /**
  * Allows the user to update their preferences
