+++ src/main/java/CLI/Pages/RecipeViewerPage.java
@@ -1,5 +1,7 @@
 package main.java.CLI.Pages;
 
+import main.java.CLI.Pages.Commands.AddToRecipeBookCommand;
+import main.java.CLI.Pages.Commands.Command;
 import main.java.CLI.Pages.Commands.GetNewRecipeCommand;
 import main.java.CLI.Pages.Commands.RateRecipeCommand;
 
@@ -10,6 +12,7 @@ public class RecipeViewerPage extends Page {
         Command[] commands = {
                 new GetNewRecipeCommand(),
                 new RateRecipeCommand(),
+                new AddToRecipeBookCommand(),
                 new GoBackCommand()
         };
         setCommands(commands);
