+++ src/main/java/CLI/Pages/SignedInPage.java
@@ -1,6 +1,7 @@
 package main.java.CLI.Pages;
 
 import main.java.CLI.CommandLineInterface;
+import main.java.CLI.Pages.Commands.Command;
 import main.java.CLI.Pages.Commands.UpdatePreferencesCommand;
 import main.java.CLI.Pages.Commands.ViewHighestRatedRecipesCommand;
