+++ src/main/java/CLI/Pages/SignedOutPage.java
@@ -1,6 +1,7 @@
 package main.java.CLI.Pages;
 
 import main.java.CLI.CommandLineInterface;
+import main.java.CLI.Pages.Commands.Command;
 import main.java.CLI.Pages.Commands.CreateAccountCommand;
 import main.java.Utilities.UserManager;
